changelog
before varnish 4.0.1 varnish VMOD ABI was allowed to change between
minor varnish revision.

https://github.com/varnishcache/varnish-cache/blob/master/doc/changes.rst#changes-from-400-to-401-2014-06-24

So we include the whole varnish version in the package name.
(that the way Oxalide does)

Now the VMOD_ABI is defined does'n change so much.
So we include only the VMOD_ABI in the debian revision.

VMOD_ABI of varnish 4.1.3